created a new file engine.py and moved all engine classes to the new file.
